refactor(driver): extract step_builder + helpers to driver/step_builder.rs (refactor 2.1l)
Move step_builder (the builder-mode per-step entry point), recover_builder_l2_nonce,
collect_reverted_user_transactions, inject_held_l2_txs, and reinject_pending_transactions
from driver/mod.rs to driver/step_builder.rs. These five methods form the builder's
block-building and pool-management layer.

Also cleans up orphaned doc comments left behind from earlier extractions (verify,
clear_internal_state, protocol_txs, build, rewind doc blocks that were stranded in
mod.rs without their corresponding method bodies).

mod.rs: 2041 -> 971 lines. step_builder.rs: 1069 lines. Zero behavioral change.
608/608 tests pass, clippy clean.
